% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_cgm.R
\name{bart_cgm}
\alias{bart_cgm}
\title{Run the BART algorithm for supervised learning with the basic CGM priors and Sparse option + SV}
\usage{
bart_cgm(
  y_train,
  X_train,
  X_test = NULL,
  num_burnin = 1000L,
  num_mcmc = 1000L,
  num_thin = 1L,
  general_params = list(),
  mean_forest_params = list(),
  mean_prior = "cgm",
  SV = FALSE,
  sv_priors_list = list(priormu = c(0, 1), priorphi = c(10, 3), priorsigma = c(0.5, 0.5))
)
}
\arguments{
\item{y_train}{Outcome to be modeled by the ensemble.}

\item{X_train}{Covariates used to split trees in the ensemble. May be provided either as a dataframe or a matrix.
Matrix covariates will be assumed to be all numeric. Covariates passed as a dataframe will be
preprocessed based on the variable types (e.g. categorical columns stored as unordered factors will be one-hot encoded,
categorical columns stored as ordered factors will passed as integers to the core algorithm, along with the metadata
that the column is ordered categorical).}

\item{X_test}{(Optional) Test set of covariates used to define "out of sample" evaluation data.
May be provided either as a dataframe or a matrix, but the format of \code{X_test} must be consistent with
that of \code{X_train}.}

\item{num_burnin}{Number of "burn-in" iterations of the MCMC sampler. Default: 1000L.}

\item{num_mcmc}{Number of "retained" iterations of the MCMC sampler. Default: 1000L.}

\item{num_thin}{}

\item{general_params}{}

\item{mean_forest_params}{}

\item{mean_prior}{}

\item{SV}{}

\item{sv_priors_list}{}
}
\description{
Run the BART algorithm for supervised learning with the basic CGM priors and Sparse option + SV
}
