% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_probability_prior.R
\name{draw_alpha_update}
\alias{draw_alpha_update}
\title{Title}
\usage{
draw_alpha_update(alpha, lpv, a, b, rho)
}
\arguments{
\item{alpha}{}

\item{lpv}{}

\item{a}{}

\item{b}{}

\item{rho}{}
}
\description{
Title
}
