% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_probability_prior.R
\name{draw_minessota_split}
\alias{draw_minessota_split}
\title{Title}
\usage{
draw_minessota_split(nv, m_i, lag_index, sigma_ols, lambda_1, lambda_2, rng)
}
\arguments{
\item{nv}{Vector with the number of splits for each variable}

\item{m_i}{Endogenous variable index}

\item{lag_index}{index for choosing the appropiate weight in the dirichilet}

\item{sigma_ols}{calibrated weight for the cross-lags}

\item{lambda_1}{shrinkage parameter for i = j}

\item{lambda_2}{shrinkage parameter for i != j}

\item{rng}{}
}
\value{
List of probabilities according to the Minessota criteria.
}
\description{
Title
}
