% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_probability_prior.R
\name{draw_theta_update}
\alias{draw_theta_update}
\title{Title}
\usage{
draw_theta_update(theta, lpv, a, b, rho)
}
\arguments{
\item{theta}{}

\item{lpv}{}

\item{a}{}

\item{b}{}

\item{rho}{}
}
\description{
Title
}
