% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv_bart.R
\name{fsv_mbart}
\alias{fsv_mbart}
\title{Title}
\usage{
fsv_mbart(
  y_raw,
  lags = 1,
  fhorz = 1,
  bart_prior = "CGM",
  sv = "none",
  keep_samples = FALSE,
  num_burnin = 100,
  num_mcmc = 100,
  num_thin = 1,
  params = ls()
)
}
\arguments{
\item{y_raw}{}

\item{lags}{}

\item{fhorz}{}

\item{bart_prior}{}

\item{sv}{}

\item{keep_samples}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{params}{}
}
\description{
Title
}
