% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv_varbart.r
\name{fsv_varbart}
\alias{fsv_varbart}
\title{Title}
\usage{
fsv_varbart(
  Yraw,
  p,
  fhorz,
  sv = "none",
  bart_prior = "standard",
  a_dart = 0.5,
  b_dart = 1,
  rho_dart = NULL,
  lambda_1 = 1,
  lambda_2 = 1,
  lambda_update = F,
  cutpoint_grid_size = 100,
  tau_init = NULL,
  alpha = 0.95,
  beta = 2,
  min_samples_leaf = 5,
  output_dimension = 1,
  is_leaf_constant = T,
  nu = 3,
  leaf_model = 0,
  lambda = NULL,
  a_leaf = 3,
  b_leaf = NULL,
  sigma2_init = NULL,
  num_trees = 250,
  num_burnin = 250,
  num_mcmc = 100,
  num_thin = 1,
  random_seed = -1
)
}
\arguments{
\item{Yraw}{}

\item{p}{}

\item{fhorz}{}

\item{sv}{}

\item{a_dart}{}

\item{b_dart}{}

\item{rho_dart}{}

\item{lambda_1}{}

\item{lambda_2}{}

\item{cutpoint_grid_size}{}

\item{tau_init}{}

\item{alpha}{}

\item{beta}{}

\item{min_samples_leaf}{}

\item{output_dimension}{}

\item{is_leaf_constant}{}

\item{nu}{}

\item{leaf_model}{}

\item{lambda}{}

\item{a_leaf}{}

\item{b_leaf}{}

\item{sigma2_init}{}

\item{num_trees}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{random_seed}{}

\item{Variable_Split_Prior}{}

\item{Theta_Update}{}

\item{Lambda_Update}{}

\item{q}{}

\item{num_gfr}{}
}
\description{
Title
}
