% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_utils.r
\name{get_hs}
\alias{get_hs}
\title{Title}
\usage{
get_hs(bdraw, lambda.hs, nu.hs, tau.hs, zeta.hs)
}
\arguments{
\item{bdraw}{}

\item{lambda.hs}{}

\item{nu.hs}{}

\item{tau.hs}{}

\item{zeta.hs}{}
}
\description{
Title
}
