% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_utils.R
\name{get_Lambda}
\alias{get_Lambda}
\title{Title}
\usage{
get_Lambda(eps, fac, S, pr, m, q, id.fac)
}
\arguments{
\item{eps}{}

\item{fac}{}

\item{S}{}

\item{pr}{}

\item{m}{}

\item{q}{}

\item{id.fac}{}
}
\description{
Title
}
