% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv_bart.R
\name{predict.fsv_mbart}
\alias{predict.fsv_mbart}
\title{(Recursive) Predict from a sampled VAR-BART model on new data}
\usage{
\method{predict}{fsv_mbart}(fsv_mbart, Y_test, n_ahead = 1L)
}
\arguments{
\item{fsv_mbart}{Object of type \code{fsv_mbart} containing draws of a regression forest and associated sampling outputs.}

\item{Y_test}{Data matrix of observed values for computation of log predictive
likelihood. Each of \code{ncol(fsv_mbart$Yraw)} columns is assumed to contain a
single time-series of length \code{length(ahead)}.}

\item{n_ahead}{Integer vector, indicating the number of steps ahead at which to predict.}
}
\description{
(Recursive) Predict from a sampled VAR-BART model on new data
}
