% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_split.R
\name{sample_alpha_one_iteration}
\alias{sample_alpha_one_iteration}
\title{Sample one iteration of the scale of the Dirichlet Distribution}
\usage{
sample_alpha_one_iteration(log_prob_vector, a, b, rho, rng)
}
\arguments{
\item{log_prob_vector}{}

\item{a}{}

\item{b}{}

\item{rho}{}

\item{rng}{C++ random number generator}

\item{alpha}{}
}
\description{
Sample one iteration of the scale of the Dirichlet Distribution
}
