% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_split.R
\name{sample_dart_splits_one_iteration}
\alias{sample_dart_splits_one_iteration}
\title{Sample one iteration of the split probabilities following Linero(2018)}
\usage{
sample_dart_splits_one_iteration(variable_count_splits, alpha, rng)
}
\arguments{
\item{variable_count_splits}{number of splits for each variable}

\item{alpha}{scale of the dirichilet distribtuion}

\item{rng}{C++ random number generator}
}
\description{
Sample one iteration of the split probabilities following Linero(2018)
}
