% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbart.R
\name{varbart}
\alias{varbart}
\title{Title}
\usage{
varbart(
  data,
  Y_test = NULL,
  n_ahead = NULL,
  lags = 1L,
  mean_prior = "cgm",
  variance_prior = "const",
  cov_shrinkage = "hs",
  num_factors = NULL,
  num_warmstart = 0L,
  num_burnin = 100L,
  num_mcmc = 100L,
  num_thin = 1L,
  general_params = list(),
  mean_forest_params = list()
)
}
\arguments{
\item{data}{}

\item{Y_test}{}

\item{n_ahead}{}

\item{lags}{}

\item{mean_prior}{}

\item{variance_prior}{}

\item{cov_shrinkage}{}

\item{num_factors}{}

\item{num_warmstart}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{num_thin}{}

\item{general_params}{}

\item{mean_forest_params}{}
}
\description{
Title
}
