% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv_bart.R
\name{fsv_mbart}
\alias{fsv_mbart}
\title{Run the Multivariate BART Algorithm for Supervised Learning.}
\usage{
fsv_mbart(
  data,
  lags = 1L,
  bart_prior = "CGM",
  SV = FALSE,
  num_burnin = 100L,
  num_mcmc = 100L,
  num_thin = 1L,
  keep_burnin = FALSE,
  params = ls()
)
}
\arguments{
\item{data}{Outcome to be modeled by the ensemble. This is a TxM matriz.}

\item{lags}{Number of Lags used in the VAR(p) system.}

\item{bart_prior}{Which Prior we are going to use in the BART: CGM, DART or Minn.}

\item{SV}{Stochastic Volatility Flag.}

\item{num_burnin}{Number of "burn-in" iterations of the MCMC sampler. Default: 100.}

\item{num_mcmc}{Number of "retained" iterations of the MCMC sampler. Default: 100.}

\item{num_thin}{}

\item{keep_burnin}{Whether or not "burn-in" samples should be stored. Default FALSE.}

\item{params}{}
}
\value{
List of sampling outputs and a wrapper around the sampled forests (which can be used for in-memory prediction on new data).
}
\description{
Run the Multivariate BART Algorithm for Supervised Learning.
}
