% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsv_varbart.r
\name{fsv_varbart}
\alias{fsv_varbart}
\title{Title}
\usage{
fsv_varbart(
  Y_train,
  X_train,
  Yraw,
  p,
  fhorz,
  Heteroskedacity_Model = "None",
  Variable_Split_Prior = "Uniform",
  Theta_Update = F,
  cutpoint_grid_size = 100,
  tau_init = NULL,
  alpha = 0.95,
  beta = 2,
  min_samples_leaf = 5,
  output_dimension = 1,
  is_leaf_constant = T,
  nu = 3,
  leaf_model = 0,
  lambda = NULL,
  a_leaf = 3,
  b_leaf = NULL,
  q = 0.9,
  sigma2_init = NULL,
  num_trees = 250,
  num_gfr = 0,
  num_burnin = 250,
  num_mcmc = 100,
  random_seed = -1
)
}
\arguments{
\item{Y_train}{}

\item{X_train}{}

\item{Yraw}{}

\item{p}{}

\item{fhorz}{}

\item{Heteroskedacity_Model}{}

\item{Variable_Split_Prior}{}

\item{Theta_Update}{}

\item{cutpoint_grid_size}{}

\item{tau_init}{}

\item{alpha}{}

\item{beta}{}

\item{min_samples_leaf}{}

\item{output_dimension}{}

\item{is_leaf_constant}{}

\item{nu}{}

\item{leaf_model}{}

\item{lambda}{}

\item{a_leaf}{}

\item{b_leaf}{}

\item{q}{}

\item{sigma2_init}{}

\item{num_trees}{}

\item{num_gfr}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{random_seed}{}
}
\description{
Title
}
