% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbart.R
\name{varbart}
\alias{varbart}
\title{Title}
\usage{
varbart(
  y_train,
  X_train,
  sv = FALSE,
  W_train = NULL,
  group_ids_train = NULL,
  rfx_basis_train = NULL,
  X_test = NULL,
  W_test = NULL,
  group_ids_test = NULL,
  rfx_basis_test = NULL,
  ordered_cat_vars = NULL,
  unordered_cat_vars = NULL,
  cutpoint_grid_size = 100,
  tau_init = NULL,
  alpha = 0.95,
  beta = 2,
  min_samples_leaf = 5,
  output_dimension = 1,
  is_leaf_constant = T,
  leaf_regression = F,
  nu = 3,
  lambda = NULL,
  a_leaf = 3,
  b_leaf = NULL,
  q = 0.9,
  sigma2_init = NULL,
  num_trees = 250,
  num_gfr = 0,
  num_burnin = 0,
  num_mcmc = 100,
  random_seed = -1,
  keep_burnin = F
)
}
\arguments{
\item{y_train}{}

\item{X_train}{}

\item{sv}{}

\item{W_train}{}

\item{group_ids_train}{}

\item{rfx_basis_train}{}

\item{X_test}{}

\item{W_test}{}

\item{group_ids_test}{}

\item{rfx_basis_test}{}

\item{ordered_cat_vars}{}

\item{unordered_cat_vars}{}

\item{cutpoint_grid_size}{}

\item{tau_init}{}

\item{alpha}{}

\item{beta}{}

\item{min_samples_leaf}{}

\item{output_dimension}{}

\item{is_leaf_constant}{}

\item{leaf_regression}{}

\item{nu}{}

\item{lambda}{}

\item{a_leaf}{}

\item{b_leaf}{}

\item{q}{}

\item{sigma2_init}{}

\item{num_trees}{}

\item{num_gfr}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{random_seed}{}

\item{keep_burnin}{}
}
\description{
Title
}
