% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varbart.R
\name{varbart}
\alias{varbart}
\title{Title}
\usage{
varbart(
  Yraw,
  p,
  fhorz,
  Heteroskedacity_Model = "homo",
  SUR = FALSE,
  Variable_Split_Prior = "uniform",
  Theta_Update = T,
  a_dart = 0.5,
  b_dart = 1,
  rho_dart = NULL,
  Lambda_Update = F,
  lambda_1 = NULL,
  lambda_2 = NULL,
  a_minn = 0.5,
  b_minn = 1,
  cutpoint_grid_size = 100,
  tau_init = NULL,
  alpha = 0.95,
  beta = 2,
  min_samples_leaf = 5,
  output_dimension = 1,
  is_leaf_constant = T,
  nu = 3,
  leaf_model = 0,
  lambda = NULL,
  a_leaf = 3,
  b_leaf = NULL,
  q = 0.9,
  sigma2_init = NULL,
  num_trees = 250,
  num_gfr = 0,
  num_burnin = 250,
  num_mcmc = 100,
  random_seed = -1,
  keep_burnin = F
)
}
\arguments{
\item{Yraw}{}

\item{p}{}

\item{fhorz}{}

\item{Heteroskedacity_Model}{}

\item{SUR}{}

\item{Variable_Split_Prior}{}

\item{Theta_Update}{}

\item{a_dart}{}

\item{b_dart}{}

\item{rho_dart}{}

\item{Lambda_Update}{}

\item{lambda_1}{}

\item{lambda_2}{}

\item{a_minn}{}

\item{b_minn}{}

\item{cutpoint_grid_size}{}

\item{tau_init}{}

\item{alpha}{}

\item{beta}{}

\item{min_samples_leaf}{}

\item{output_dimension}{}

\item{is_leaf_constant}{}

\item{nu}{}

\item{leaf_model}{}

\item{lambda}{}

\item{a_leaf}{}

\item{b_leaf}{}

\item{q}{}

\item{sigma2_init}{}

\item{num_trees}{}

\item{num_gfr}{}

\item{num_burnin}{}

\item{num_mcmc}{}

\item{random_seed}{}

\item{keep_burnin}{}
}
\description{
Title
}
